+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+* Add `test_cli.py` and `test_config.py`.
+
 ## HardPy 0.6.0
 
 In HardPy, compared to version 0.5.0 and less, the startup principle has changed.